Add canAcceptConnection(direction) method to INetworkNode to allow checking if a node can accept a visual network connection from the provided direction.  Defaults to the result of canConduct() for backward compatability, which is the proper result for most nodes.  This allows separation of visual connection logic from network conductivity logic to support blocks like the Relay and the Wireless Transmitter.
Move hasBlockingMicroblock() and isBlockingMicroblock() to TileNode make
them static and use them in canAcceptConnection() for visual connectivity.
The Relay and Wireless Transmitter both need to accept visual network
connections based on whether they are blocked by microblocks.  Because
they are descended from TileNode they would need to call these functions
statically like BlockCable does to decide on visual connectivity.  With
the addition of canAcceptConnection() BlockCable can use that to decide on
connections so there is no need for these calls to be static.  These calls
are useful for all node-based tile entities.